[vulkan] Change curFrame from size_t to uint32_t
I very much doubt that even 4294967295 frames will be needed, let alone
more. Saves me having to use %zd or %zu all over the place.

+++ include/vid_vulkan.h
@@ -61,7 +61,7 @@ typedef struct vulkan_ctx_s {
 
 	VkCommandPool cmdpool;
 	struct qfv_stagebuf_s *staging;
-	size_t      curFrame;
+	uint32_t    curFrame;
 	vulkan_frameset_t frames;
 	qfv_renderpassset_t renderPasses;
 	struct qfv_renderpass_s *output_renderpass;
